MOODI-87 Oodista kurssilta poistettujen opiskelijoiden tulee poistua myös Moodlesta - Review correction: now using test props instead of weird setter on MoodleCourseBuilder.
