+++ src/itest/resources/fixtures/oodi/course-realisation-itest.json
