+++ src/main/java/fi/helsinki/moodi/service/importing/MoodleCourseBuilder.java
@@ -24,6 +24,7 @@ import fi.helsinki.moodi.integration.oodi.OodiLocalizedValue;
 import fi.helsinki.moodi.service.util.MapperService;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -35,6 +36,8 @@ public class MoodleCourseBuilder {
 
     public static final String MOODLE_COURSE_ID_PREFIX = "oodi_";
     public static final int DEFAULT_NUMBER_OF_SECTIONS = 7;
+
+    @Value("${test.MoodleCourseBuilder.courseVisibility:false}")
     private boolean courseVisibility = false;
 
     private final MapperService mapperService;
@@ -61,10 +64,6 @@ public class MoodleCourseBuilder {
             courseVisibility,
             DEFAULT_NUMBER_OF_SECTIONS
         );
-    }
-
-    public void useOnlyInTestsToSetCourseVisibility(boolean courseVisibility) {
-        this.courseVisibility = courseVisibility;
     }
 
     private String getDescription(OodiCourseUnitRealisation oodiCourseUnitRealisation, String language) {
